Convert subscription turnip feature to rspec
